Reorganize transaction traits
