+++ crates/cosmos/cosmos-client-components/src/impls/transaction/encode_tx.rs
@@ -1,7 +1,7 @@
 use cgp_core::CanRaiseError;
 use hermes_relayer_components::chain::traits::types::chain_id::HasChainId;
 use hermes_relayer_components::chain::traits::types::message::HasMessageType;
-use hermes_relayer_components::transaction::traits::components::tx_encoder::TxEncoder;
+use hermes_relayer_components::transaction::traits::encode_tx::TxEncoder;
 use hermes_relayer_components::transaction::traits::types::fee::HasFeeType;
 use hermes_relayer_components::transaction::traits::types::nonce::HasNonceType;
 use hermes_relayer_components::transaction::traits::types::signer::HasSignerType;
