+++ crates/cosmos/cosmos-client-components/src/impls/transaction/query_tx_response.rs
@@ -1,5 +1,5 @@
 use cgp_core::CanRaiseError;
-use hermes_relayer_components::transaction::traits::components::tx_response_querier::TxResponseQuerier;
+use hermes_relayer_components::transaction::traits::query_tx_response::TxResponseQuerier;
 use hermes_relayer_components::transaction::traits::types::tx_hash::HasTransactionHashType;
 use hermes_relayer_components::transaction::traits::types::tx_response::HasTxResponseType;
 use ibc_relayer::chain::cosmos::query::tx::query_tx_response;
