+++ crates/cosmos/cosmos-client-components/src/impls/types/transaction.rs
@@ -1,5 +1,5 @@
 use cgp_core::Async;
-use hermes_relayer_components::transaction::traits::nonce::guard::ProvideNonceGuard;
+use hermes_relayer_components::transaction::traits::nonce::nonce_guard::ProvideNonceGuard;
 use hermes_relayer_components::transaction::traits::types::fee::ProvideFeeType;
 use hermes_relayer_components::transaction::traits::types::nonce::{
     HasNonceType, ProvideNonceType,
