+++ crates/cosmos/cosmos-relayer/src/impls/transaction/fields.rs
@@ -3,9 +3,9 @@ use hermes_cosmos_client_components::types::nonce_guard::NonceGuard;
 use hermes_relayer_components::chain::traits::types::chain_id::ChainIdGetter;
 use hermes_relayer_components::runtime::traits::mutex::MutexGuardOf;
 use hermes_relayer_components::runtime::traits::runtime::ProvideRuntime;
-use hermes_relayer_components::transaction::traits::fee::FeeForSimulationGetter;
-use hermes_relayer_components::transaction::traits::nonce::mutex::ProvideMutexForNonceAllocation;
-use hermes_relayer_components::transaction::traits::signer::DefaultSignerGetter;
+use hermes_relayer_components::transaction::traits::default_signer::DefaultSignerGetter;
+use hermes_relayer_components::transaction::traits::nonce::nonce_mutex::ProvideMutexForNonceAllocation;
+use hermes_relayer_components::transaction::traits::simulation_fee::FeeForSimulationGetter;
 use hermes_relayer_runtime::types::runtime::HermesRuntime;
 use ibc_proto::cosmos::tx::v1beta1::Fee;
 use ibc_relayer::chain::cosmos::types::account::Account;
