+++ crates/relayer/relayer-components/src/transaction/components/limit_tx_size.rs
@@ -10,7 +10,7 @@ use core::marker::PhantomData;
 use cgp_core::CanRaiseError;
 
 use crate::chain::traits::types::message::HasMessageType;
-use crate::transaction::traits::components::tx_encoder::TxEncoder;
+use crate::transaction::traits::encode_tx::TxEncoder;
 use crate::transaction::traits::types::fee::HasFeeType;
 use crate::transaction::traits::types::nonce::HasNonceType;
 use crate::transaction::traits::types::signer::HasSignerType;
