+++ crates/relayer/relayer-components/src/transaction/components/send_messages_with_default_signer.rs
@@ -1,8 +1,8 @@
 use alloc::vec::Vec;
 
 use crate::chain::traits::send_message::MessageSender;
-use crate::transaction::traits::components::send_messages_with_signer::CanSendMessagesWithSigner;
-use crate::transaction::traits::signer::HasDefaultSigner;
+use crate::transaction::traits::default_signer::HasDefaultSigner;
+use crate::transaction::traits::send_messages_with_signer::CanSendMessagesWithSigner;
 
 pub struct SendMessagesWithDefaultSigner;
