--- crates/relayer/relayer-components/src/transaction/traits/components/mod.rs
@@ -1,7 +1,0 @@
-pub mod send_messages_with_signer;
-pub mod send_messages_with_signer_and_nonce;
-pub mod tx_encoder;
-pub mod tx_fee_estimater;
-pub mod tx_response_poller;
-pub mod tx_response_querier;
-pub mod tx_submitter;
