+++ crates/relayer/relayer-components/src/transaction/traits/mod.rs
@@ -1,7 +1,13 @@
-pub mod components;
-pub mod event;
-pub mod fee;
+pub mod default_signer;
+pub mod encode_tx;
+pub mod estimate_tx_fee;
 pub mod logs;
 pub mod nonce;
-pub mod signer;
+pub mod parse_events;
+pub mod poll_tx_response;
+pub mod query_tx_response;
+pub mod send_messages_with_signer;
+pub mod send_messages_with_signer_and_nonce;
+pub mod simulation_fee;
+pub mod submit_tx;
 pub mod types;
