+++ crates/relayer/relayer-components/src/transaction/traits/nonce/mod.rs
@@ -1,4 +1,4 @@
 pub mod allocate_nonce;
-pub mod guard;
-pub mod mutex;
+pub mod nonce_guard;
+pub mod nonce_mutex;
 pub mod query_nonce;
