+++ crates/sovereign/sovereign-client-components/src/sovereign/impls/transaction/query_tx_response.rs
@@ -1,7 +1,7 @@
 use core::ops::Range;
 
 use cgp_core::CanRaiseError;
-use hermes_relayer_components::transaction::traits::components::tx_response_querier::TxResponseQuerier;
+use hermes_relayer_components::transaction::traits::query_tx_response::TxResponseQuerier;
 use hermes_relayer_components::transaction::traits::types::tx_hash::HasTransactionHashType;
 use hermes_relayer_components::transaction::traits::types::tx_response::HasTxResponseType;
 use hex::ToHex;
