+++ crates/sovereign/sovereign-integration-tests/tests/bootstrap.rs
@@ -9,9 +9,9 @@ use hermes_celestia_integration_tests::contexts::bootstrap::CelestiaBootstrap;
 use hermes_celestia_test_components::bootstrap::traits::bootstrap_bridge::CanBootstrapBridge;
 use hermes_cosmos_relayer::contexts::builder::CosmosBuilder;
 use hermes_cosmos_relayer::types::error::Error;
-use hermes_relayer_components::transaction::traits::components::tx_response_poller::CanPollTxResponse;
-use hermes_relayer_components::transaction::traits::components::tx_response_querier::CanQueryTxResponse;
-use hermes_relayer_components::transaction::traits::event::CanParseTxResponseAsEvents;
+use hermes_relayer_components::transaction::traits::parse_events::CanParseTxResponseAsEvents;
+use hermes_relayer_components::transaction::traits::poll_tx_response::CanPollTxResponse;
+use hermes_relayer_components::transaction::traits::query_tx_response::CanQueryTxResponse;
 use hermes_relayer_runtime::types::runtime::HermesRuntime;
 use hermes_sovereign_client_components::sovereign::traits::rollup::publish_batch::CanPublishTransactionBatch;
 use hermes_sovereign_client_components::sovereign::types::message::SovereignMessage;
